pyproject.toml
[tool.poetry]
name = "text-2-image"
version = "0.1.0"
description = ""
authors = ["ash.blade <ashen.blade.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "src/text_to_image"}]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"


[tool.poetry.group.dev.dependencies]
mypy = "^0.981"
pytest = "^7.1.3"
flake8 = "^5.0.4"
black = "^22.10.0"

[tool.pytest.ini_options]
pythonpath = ['src']

[tool.mypy]
exclude = [
    'venv',
    'tests'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
